Ab0utME.md
Currently Biomedical Engineering student, interested in Healthcare, Engineering, Internet, Economy and AI. With several years of professional experience, working with technical services and support in the area of sales, costumer service, online research and operation of software for systems. Bilingual and self-taught person with teamwork and leadership spirit. Microsoft office specialist and taking courses to programming AI. Organized and liable person that is still learning